Refactor: migrate to typescript and refactor server
