+++ server/package.json
@@ -1,11 +1,11 @@
 {
   "name": "server",
-  "version": "2.1.0",
+  "version": "3.0.0",
   "main": "index.js",
   "license": "MIT",
   "type": "module",
   "dependencies": {
-    "axios": "^1.5.1",
+    "axios": "^1.6.2",
     "cheerio": "^1.0.0-rc.12",
     "cors": "^2.8.5",
     "dotenv": "^16.3.1",
@@ -22,7 +22,12 @@
     "server:dev": "nodemon --exec ts-node app.ts"
   },
   "devDependencies": {
+    "@types/cors": "^2.8.16",
+    "@types/express": "^4.17.21",
+    "@types/node": "^20.9.2",
+    "@types/node-cron": "^3.0.11",
     "@types/nodemailer": "^6.4.11",
+    "@types/xml2js": "^0.4.14",
     "nodemon": "^3.0.1",
     "ts-node": "^10.9.1",
     "typescript": "^5.2.2"
