+++ server/types/DepartmentType.ts
@@ -1,0 +1,7 @@
+export interface DepartmentType {
+  code: string;
+  name: string;
+  url: string;
+  boards: number[];
+  board_names: string[];
+}
