+++ server/types/MessageType.ts
@@ -1,0 +1,17 @@
+export interface MessageContentType {
+  [postIndex: string]: {
+    title: string;
+    link: string;
+    pubDate: string;
+    images: string[];
+  };
+}
+
+export interface MessageType {
+  latestPostIndex?: number;
+  message?: MessageContentType;
+}
+
+export interface MessagesType {
+  [board_name: string]: MessageType;
+}
