+++ server/types/UserType.ts
@@ -1,0 +1,10 @@
+export interface UserType {
+  email: string;
+  latest_post_indexes: number[];
+  department_code: string;
+  subscribe_time: Date;
+}
+
+export interface UsersType {
+  users: UserType[];
+}
